Remove or move several `TypeAlias` declarations
