+++ stdlib/lib2to3/pgen2/__init__.pyi
@@ -1,0 +1,7 @@
+from collections.abc import Callable
+from lib2to3.pgen2.grammar import Grammar
+from lib2to3.pytree import _RawNode
+from typing import Any
+from typing_extensions import TypeAlias
+
+_Convert: TypeAlias = Callable[[Grammar, _RawNode], Any]
