+++ stdlib/lib2to3/pgen2/driver.pyi
@@ -1,7 +1,8 @@
 from _typeshed import StrPath
 from collections.abc import Iterable
+from lib2to3.pgen2 import _Convert
 from lib2to3.pgen2.grammar import Grammar
-from lib2to3.pytree import _NL, _Convert
+from lib2to3.pytree import _NL
 from logging import Logger
 from typing import IO, Any
