+++ stdlib/lib2to3/pytree.pyi
@@ -1,5 +1,5 @@
 from _typeshed import Self
-from collections.abc import Callable, Iterator
+from collections.abc import Iterator
 from lib2to3.pgen2.grammar import Grammar
 from typing import Any
 from typing_extensions import TypeAlias
@@ -8,9 +8,6 @@ _NL: TypeAlias = Node | Leaf
 _Context: TypeAlias = tuple[str, int, int]
 _Results: TypeAlias = dict[str, _NL]
 _RawNode: TypeAlias = tuple[int, str, _Context, list[_NL] | None]
-# This alias isn't used in this file,
-# but is imported in other lib2to3 submodules
-_Convert: TypeAlias = Callable[[Grammar, _RawNode], Any]
 
 HUGE: int
