+++ stubs/SQLAlchemy/sqlalchemy/ext/mypy/infer.pyi
@@ -7,7 +7,6 @@ _Expression: TypeAlias = Any  # mypy.nodes.Expression
 _RefExpr: TypeAlias = Any  # mypy.nodes.RefExpr
 _TypeInfo: TypeAlias = Any  # mypy.nodes.TypeInfo
 _Var: TypeAlias = Any  # mypy.nodes.Var
-_StrExpr: TypeAlias = Any  # mypy.nodes.StrExpr
 _SemanticAnalyzerPluginInterface: TypeAlias = Any  # mypy.plugin.SemanticAnalyzerPluginInterface
 _ProperType: TypeAlias = Any  # mypy.types.ProperType
